+++ test/config_test.rb
@@ -157,11 +157,13 @@ class ConfigTest < Minitest::Test
 
   def test_content_with_broken_yaml
     exn = assert_raises Runners::Config::BrokenYAML do
-      Runners::Config.new(path: Pathname(FILE_NAME), raw_content: "@").content
+      Runners::Config.new(path: Pathname(FILE_NAME), raw_content: "\n  @\n").content
     end
-    assert_equal "`sider.yml` is broken at line 1 and column 1", exn.message
+    assert_equal "`sider.yml` is broken at line 2 and column 3", exn.message
     assert_equal "sider.yml", exn.path_name
-    assert_equal "@", exn.raw_content
+    assert_equal "\n  @\n", exn.raw_content
+    assert_equal 2, exn.line
+    assert_equal 3, exn.column
   end
 
   def test_path_name